save-session sidenav updated
