--- ui/src/app/components/save-session-form/save-session-form.component.html
@@ -1,45 +1,0 @@
-<div mat-dialog-content>
-  <form [formGroup]="saveSessionFrom" class="save-session-form">
-    <h3>Save Session</h3>
-
-    <mat-form-field class="full-width" appearance="outline">
-      <mat-label>Session Name</mat-label>
-      <input matInput placeholder="mysession" type="text" formControlName="SessionName" />
-    </mat-form-field>
-    <br />
-    <mat-form-field class="full-width" appearance="outline">
-      <mat-label>Editor Name</mat-label>
-      <input matInput placeholder="editor name" type="text" formControlName="EditorName" />
-    </mat-form-field>
-
-    <br />
-    <mat-form-field class="full-width" appearance="outline">
-      <mat-label>Database Name</mat-label>
-      <input matInput placeholder="Bikestore" type="text" formControlName="DatabaseName" />
-    </mat-form-field>
-    <br />
-    <mat-form-field class="full-width" appearance="outline">
-      <mat-label>Notes</mat-label>
-      <textarea
-        rows="2"
-        matInput
-        placeholder="added new index"
-        type="text"
-        formControlName="Notes"
-      ></textarea>
-    </mat-form-field>
-    <br />
-  </form>
-</div>
-<div mat-dialog-actions>
-  <button mat-raised-button color="primary" mat-dialog-close>Close</button>
-  <button
-    mat-raised-button
-    (click)="saveSession()"
-    type="submit"
-    color="primary"
-    [disabled]="!saveSessionFrom.valid"
-  >
-    Save
-  </button>
-</div>
