--- ui/src/app/components/save-session-form/save-session-form.component.scss
@@ -1,9 +1,0 @@
-.save-session-form {
-  min-width: 150px;
-  max-width: 500px;
-  width: 100%;
-}
-
-.full-width {
-  width: 100%;
-}
