+++ ui/src/app/components/sidenav-save-session/sidenav-save-session.component.ts
@@ -1,0 +1,62 @@
+import { Component, OnInit } from '@angular/core'
+import { FormControl, FormGroup, Validators } from '@angular/forms'
+import { FetchService } from 'src/app/services/fetch/fetch.service'
+import { ISaveSessionPayload } from 'src/app/model/Session'
+import { DataService } from 'src/app/services/data/data.service'
+import { SnackbarService } from '../../services/snackbar/snackbar.service'
+import { SidenavService } from 'src/app/services/sidenav/sidenav.service'
+
+@Component({
+  selector: 'app-sidenav-save-session',
+  templateUrl: './sidenav-save-session.component.html',
+  styleUrls: ['./sidenav-save-session.component.scss'],
+})
+export class SidenavSaveSessionComponent implements OnInit {
+  errMessage: string = ''
+  constructor(
+    private fetch: FetchService,
+    private data: DataService,
+    private snack: SnackbarService,
+    private sidenav: SidenavService
+  ) {}
+  saveSessionForm: FormGroup = new FormGroup({
+    SessionName: new FormControl('', [
+      Validators.required,
+      Validators.pattern('^[a-zA-Z].{0,49}$'),
+    ]),
+    EditorName: new FormControl('', [Validators.required, Validators.pattern('^[a-zA-Z].{0,49}$')]),
+    DatabaseName: new FormControl('', [
+      Validators.required,
+      Validators.pattern('^[a-zA-Z].{0,49}$'),
+    ]),
+    Notes: new FormControl('', [Validators.required]),
+  })
+
+  saveSession() {
+    let formValue = this.saveSessionForm.value
+    let payload: ISaveSessionPayload = {
+      SessionName: formValue.SessionName,
+      EditorName: formValue.EditorName,
+      DatabaseName: formValue.DatabaseName,
+      Notes: formValue.Notes.split('\n'),
+    }
+
+    this.fetch.saveSession(payload).subscribe({
+      next: (res: any) => {
+        this.data.getAllSessions()
+        this.snack.openSnackBar('Session saved successfully', 'Close', 5000)
+      },
+      error: (err: any) => {
+        this.snack.openSnackBar(err.error, 'Close', 5000)
+      },
+    })
+    this.saveSessionForm.reset()
+    this.saveSessionForm.markAsUntouched()
+    this.closeSidenav()
+  }
+
+  ngOnInit(): void {}
+  closeSidenav(): void {
+    this.sidenav.closeSidenav()
+  }
+}
