+++ ui/src/app/components/workspace/workspace.component.html
@@ -7,7 +7,7 @@
     </div>
     <div class="header_action">
       <button mat-button (click)="openAssessment()">VIEW ASSESSMENT</button>
-      <button mat-button (click)="openSaveSessionModal()" *ngIf="!isOfflineStatus">
+      <button mat-button (click)="openSaveSessionSidenav()" *ngIf="!isOfflineStatus">
         SAVE SESSION
       </button>
       <button mat-button (click)="downloadSession()">DOWNLOAD SESSION</button>
